fix(core): move command helper out of auto imports
Keep the conversation target helper out of the generated commands entry so process-dynamic-import no longer treats it as a command module that must export apply.
