+++ packages/core/src/utils/conversation.ts
